use f32 as the cuda code to be able to compare results added debug render enabled AA in cuda_compare example

+++ examples/shadow_glamour_shot.rs
@@ -4,7 +4,7 @@ extern crate num_cpus;
 
 use std::any::Any;
 use std::error::Error;
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 use std::sync::{Arc, Mutex};
 use std::thread;
 use std::time::Instant;
@@ -130,7 +130,7 @@ fn main() -> Result<(), Box<dyn Error>> {
 
                                 color = color + World::color_at(&w_clone, &r, MAX_REFLECTION_RECURSION_DEPTH);
                             }
-                            color = color / n_samples as f64;
+                            color = color / n_samples as f32;
                         // println!("with AA    color at ({}/{}): {:?}", x, y, color);
                         } else {
                             let r = Camera::ray_for_pixel(&c_clone, x, y);
